refactor: 패키지 구조 대변경 (Controller 여정/여행 분리)
최상위 패키지: common, config, domain, view(다음 프로젝트에서 삭제)

domain 패키지에서 trip, itinerary, file(다음 프로젝트에서 삭제)로 나뉘고
각 도메인(trip, itinerary)마다 controller, entity, repository, service를 가짐
