+++ src/main/java/kr/co/fastcampus/travel/domain/trip/controller/dto/ItinerarySaveRequest.java
@@ -1,0 +1,23 @@
+package kr.co.fastcampus.travel.domain.trip.controller.dto;
+
+import java.time.LocalDateTime;
+import lombok.Builder;
+
+@Builder
+public record ItinerarySaveRequest(
+    String departure,
+    String destination,
+    LocalDateTime departureAt,
+    LocalDateTime arriveAt,
+    String accommodation,
+    LocalDateTime checkInAt,
+    LocalDateTime checkOutAt
+) {
+
+    public ItinerarySaveRequest {
+        if ((departureAt == null && arriveAt == null)
+            && (checkInAt == null && checkOutAt == null)) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
