+++ src/main/java/kr/co/fastcampus/travel/domain/trip/controller/dto/TripInfoResponse.java
@@ -1,0 +1,22 @@
+package kr.co.fastcampus.travel.domain.trip.controller.dto;
+
+import java.time.LocalDate;
+
+import kr.co.fastcampus.travel.domain.trip.entity.Trip;
+
+public record TripInfoResponse(
+    Long id,
+    String name,
+    LocalDate startAt,
+    LocalDate endAt
+) {
+
+    public TripInfoResponse(Trip trip) {
+        this(
+            trip.getId(),
+            trip.getName(),
+            trip.getStartAt(),
+            trip.getEndAt()
+        );
+    }
+}
