+++ src/main/java/kr/co/fastcampus/travel/domain/trip/repository/TripRepository.java
@@ -1,0 +1,15 @@
+package kr.co.fastcampus.travel.domain.trip.repository;
+
+import java.util.List;
+import java.util.Optional;
+import kr.co.fastcampus.travel.domain.trip.entity.Trip;
+import kr.co.fastcampus.travel.view.enums.FileType;
+
+public interface TripRepository {
+
+    List<Trip> findAll(FileType fileType);
+
+    Optional<Trip> findById(FileType fileType, Long id);
+
+    Trip save(Trip trip);
+}
