+++ travel/csv/trips.csv
@@ -1,0 +1,2 @@
+id,name,startAt,endAt
+1,이름,2010-01-01,2010-01-01
