+++ travel/json/trip/trip_1.json
@@ -1,0 +1,10 @@
+[ {
+  "id" : 1,
+  "departure" : "출발",
+  "destination" : "도착",
+  "departureAt" : "2010-01-01T01:01",
+  "arriveAt" : "2010-01-01T02:02",
+  "accommodation" : "숙박지",
+  "checkInAt" : "",
+  "checkOutAt" : ""
+} ]
