+++ travel/json/trips.json
@@ -1,0 +1,6 @@
+[ {
+  "id" : 1,
+  "name" : "이름",
+  "startAt" : "2010-01-01",
+  "endAt" : "2010-01-01"
+} ]
