+++ travel/sequence.json
@@ -1,0 +1,4 @@
+{
+  "trip_id" : 2,
+  "itinerary_id" : 2
+}
